+++ readme/RELEASE_NOTES.md
@@ -1,14 +1,14 @@
-[![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=${project.version}&color=blue)](https://search.maven.org/artifact/${project.groupId}/${project.artifactId}/${project.version}/jar) [![JavaDoc](https://img.shields.io/static/v1?label=JavaDoc&message=${project.version}&color=orange)](http://www.javadoc.io/doc/${project.groupId}/${project.artifactId}/${project.version})
+[![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=${project.version}&color=blue)](https://search.maven.org/artifact/${project.groupId}/restrict-imports-enforcer-rule/${project.version}/jar) [![JavaDoc](https://img.shields.io/static/v1?label=JavaDoc&message=${project.version}&color=orange)](http://www.javadoc.io/doc/${project.groupId}/restrict-imports-enforcer-rule/${project.version})
 
 ### Bug fixes
-* [#79](https://github.com/skuzzle/restrict-imports-enforcer-rule/issues/79) Recognize full qualified annotation use
+* [#63](https://github.com/skuzzle/restrict-imports-enforcer-rule/issues/63)  Support for ignoring classes that cannot be fixed
 
 Maven Central coordinates for this release:
 
 ```xml
 <dependency>
     <groupId>${project.groupId}</groupId>
-    <artifactId>${project.artifactId}</artifactId>
+    <artifactId>restrict-imports-enforcer-rule</artifactId>
     <version>${project.version}</version>
 </dependency>
 ```